moved  RenderActiveShape to its own file
