+++ client/src/modules/EmissionPieChart.js
@@ -4,6 +4,7 @@ import { Cell, Pie, PieChart, Sector, ResponsiveContainer } from "recharts";
 import "./EmissionPieChart.css";
 import Spinner from "./Spinner";
 import SubSectorInfo from "./SubSectorInfo";
+import RenderActiveShape from "./RenderActiveShape";
 
 /** TODO
  *  axios calls to get data from database
@@ -40,73 +41,6 @@ export default function EmissionPieChart() {
     }, 500);
   }, []);
 
-  const renderActiveShape = (props) => {
-    const RADIAN = Math.PI / 180;
-    const {
-      cx,
-      cy,
-      midAngle,
-      innerRadius,
-      outerRadius,
-      startAngle,
-      endAngle,
-      fill,
-      percent,
-    } = props;
-    const sin = Math.sin(-RADIAN * midAngle);
-    const cos = Math.cos(-RADIAN * midAngle);
-    const sx = cx + (outerRadius + 10) * cos;
-    const sy = cy + (outerRadius + 10) * sin;
-    const mx = cx + (outerRadius + 30) * cos;
-    const my = cy + (outerRadius + 30) * sin;
-    const ex = mx + (cos >= 0 ? 1 : -1) * 22;
-    const ey = my;
-    const textAnchor = cos >= 0 ? "start" : "end";
-    return (
-      <>
-        <Sector
-          cx={cx}
-          cy={cy}
-          innerRadius={innerRadius}
-          outerRadius={outerRadius}
-          startAngle={startAngle}
-          endAngle={endAngle}
-          fill={fill}
-        />
-        <Sector
-          cx={cx}
-          cy={cy}
-          startAngle={startAngle}
-          endAngle={endAngle}
-          innerRadius={outerRadius + 6}
-          outerRadius={outerRadius + 10}
-          fill={fill}
-        />
-        <path
-          d={`M${sx},${sy}L${mx},${my}L${ex},${ey}`}
-          stroke={fill}
-          fill="none"
-        />
-        <circle cx={ex} cy={ey} r={2} fill={fill} stroke="none" />
-        <text
-          x={ex + (cos >= 0 ? 1 : -1) * 10}
-          y={ey}
-          textAnchor={textAnchor}
-          fill="#333"
-        >{`${sectorName}`}</text>
-        <text
-          x={ex + (cos >= 0 ? 1 : -1) * 10}
-          y={ey}
-          dy={18}
-          textAnchor={textAnchor}
-          fill="#999"
-        >
-          {`(C02 Emissions ${(percent * 100).toFixed(2)}%)`}
-        </text>
-      </>
-    );
-  };
-
   const onPieEnter = useCallback(
     (_, index) => {
       setActiveIndex(index);
@@ -136,7 +70,7 @@ export default function EmissionPieChart() {
         <PieChart width={1000} height={500}>
           <Pie
             activeIndex={activeIndex}
-            activeShape={renderActiveShape}
+            activeShape={<RenderActiveShape sectorName={sectorName}/>}
             data={sectorData}
             dataKey="emissions"
             cx="50%"
